requirements.txt
python-telegram-bot[all]>=20.0
nltk>=3.8
aiohttp>=3.8.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0